+++ iam/variables.tf
@@ -4,12 +4,6 @@ variable "ibmcloud_api_key" {
   type        = string
 }
 
-variable "mca_graphql_server_url" {
-  default     = "https://api.cnop-int.us.nordcloudapp.com"
-  description = "The GraphQL API url that the provider will use to make requests."
-  type        = string
-}
-
 variable "region" {
   default     = "us-south"
   description = "Region to use with the IBM Cloud Terraform provider."